+++ bika/lims/monkey/Widget.py
@@ -74,11 +74,10 @@ def isVisible(self, instance, mode='view', default="visible", field=None):
             continue
         adapter_state = adapter(instance, mode, field, state)
         adapter_name = adapter.__class__.__name__
-        logger.info("IATWidgetVisibility rule {} for {}.{} ({}): {} -> {}"
-            .format(adapter_name, instance.id, field.getName(), mode, state,
-                    adapter_state))
-        if adapter_state == state:
-            continue
+        logger.info(
+            "IATWidgetVisibility: <{} for {}.{} mode:{} sort:{}> {} -> {}"
+            .format(adapter_name, instance.id, field.getName(), mode,
+                    adapter.sort, state, adapter_state))
         return adapter_state
 
     return state