+++ CHANGES.rst
@@ -23,6 +23,7 @@ Changelog
 **Fixed**
 
 - #1423 Use the value set for ui_item property when displaying ReferenceWidget
+- #1425 Fix adapter priority for widget visibility
 - #1421 Fix Search Query for Batches Listing
 - #1418 Subscriber adapters not supported in clients listing
 - #1419 Mixed permissions for transitions in client workflow